+++ packages/rpc-ipc/src/index.js
@@ -1,8 +1,0 @@
-// @flow
-
-import StreamRPC from '@mainframe/rpc-stream'
-import ipcTransport, { type PathOrConfig } from '@mainframe/transport-ipc'
-
-export default (pathOrConfig: PathOrConfig) => {
-  return new StreamRPC(ipcTransport(pathOrConfig))
-}