+++ packages/rpc-ipc/src/index.ts
@@ -1,0 +1,6 @@
+import StreamRPC from '@mainframe/rpc-stream'
+import ipcTransport, { PathOrConfig } from '@mainframe/transport-ipc'
+
+export default function createRPC<T>(pathOrConfig: PathOrConfig) {
+  return new StreamRPC(ipcTransport<T>(pathOrConfig))
+}