+++ packages/secure-file/src/index.js
@@ -1,7 +1,0 @@
-// @flow
-
-export * from './encrypt'
-export * from './fs'
-export * from './hash'
-export * from './sign'
-export * from './types'