+++ packages/transport-ws-browser/src/index.js
@@ -1,5 +1,0 @@
-// @flow
-
-import { WebSocketSubject } from 'rxjs/webSocket'
-
-export default (url: string) => new WebSocketSubject<Object>({ url })