+++ packages/transport-ws-browser/src/index.ts
@@ -1,0 +1,7 @@
+import { WebSocketSubject } from 'rxjs/webSocket'
+
+export default function createTransport<T = any>(
+  url: string,
+): WebSocketSubject<T> {
+  return new WebSocketSubject<T>({ url })
+}