+++ packages/utils-id/src/index.js
@@ -1,9 +1,0 @@
-// @flow
-
-import nanoid from 'nanoid'
-
-export opaque type ID: string = string
-
-export const idType = (value: any): ID => (value: ID)
-
-export const uniqueID = (): ID => idType(nanoid())