+++ packages/transport-electron/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v0.2.0 (not released)
 
-Require Electron v4 as peer dependency.
+- Require Electron v4 as peer dependency.
+- Add Typescript definitions.
 
 ## v0.1.0 (2018-08-17)